<DeepsightHarmonizerIcon /> component
