+++ src/app/item-popup/DeepSightHarmonizerIcon.m.scss.d.ts
@@ -1,8 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'deepsightProgress': string;
-  'deepsightProgressBar': string;
+  'deepsightHarmonizerIcon': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
