+++ src/app/item-popup/WeaponDeepsightInfo.m.scss
@@ -12,21 +12,8 @@
   :global(.objective-progress) {
     background: var(--theme-item-popup-progress-bar-bg);
   }
+}
 
-  .deepsightHarmonizableIcon {
-    flex-grow: 1;
-    max-width: fit-content;
-
-    img {
-      display: grid;
-      width: 15px;
-      height: 15px;
-      border: 1px solid #999;
-      margin-right: 4px;
-    }
-  }
-
-  .deepsightProgressBar {
-    flex-grow: 1;
-  }
+.deepsightProgressBar {
+  flex-grow: 1;
 }
